semisymmetric quartic
