+++ src/utils/roots/__init__.py
@@ -8,14 +8,14 @@ from .rootsinfo import RootsInfo
 from .grid import GridRender
 
 from .findroot import (
-    findroot, find_nearest_root, findroot_resultant, nroots
+    findroot, find_nearest_root, findroot_resultant
 )
 
 from .tangents import RootTangent
 
 from .rationalize import (
-    rationalize, rationalize_array, rationalize_bound, rationalize_quadratic_curve,
-    square_perturbation,
+    nroots, univariate_intervals, rationalize, rationalize_array, rationalize_bound,
+    rationalize_quadratic_curve, common_region_of_conics, square_perturbation,
     cancel_denominator
 )
 
@@ -24,9 +24,9 @@ __all__ = [
     'RootTernary', 'RootAlgebraicTernary', 'RootRationalTernary',
     'RootsInfo',
     'GridRender',
-    'findroot', 'find_nearest_root', 'findroot_resultant', 'nroots',
+    'findroot', 'find_nearest_root', 'findroot_resultant',
     'RootTangent',
-    'rationalize', 'rationalize_array', 'rationalize_bound', 'rationalize_quadratic_curve',
-    'square_perturbation',
+    'nroots', 'univariate_intervals', 'rationalize', 'rationalize_array', 'rationalize_bound',
+    'rationalize_quadratic_curve', 'common_region_of_conics', 'square_perturbation',
     'cancel_denominator'
 ]
